feat: activate reference tests for geometry generation and update geometry
Since we merged the FABulous 2.0 dev branch in main, we can finally
integrate the geometry generation in the tests.
In previous FABulous versions, the geometry generation was non deterministic results,
due to the use of sets. So we couldn't check for matching references.

Rebuild and update the geometry with FABulous 2.0 beta 3.
This should not change anymore for the default fabric.

Signed-off-by: Jonas K. <j0nask@proton.me>